Update on remote and local computers separately.

+++ automlsa2/helper_functions.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# -*- coding: utf-8 -*-
 from __future__ import print_function
 import os
 import json
@@ -8,36 +9,44 @@ import logging
 from typing import List
 from hashlib import blake2b
 from signal import signal, SIGPIPE, SIGINT, SIG_DFL
+
 signal(SIGPIPE, SIG_DFL)
 signal(SIGINT, SIG_DFL)
 
 SUFFIXES = ['.nsq', '.nin', '.nhr', '.nto', '.not', '.ndb', '.ntf']
 
 
-def remove_intermediates(runid: str, intermediates: List[str], protect: bool)\
-        -> None:
+def remove_intermediates(runid: str, intermediates: List[str], protect: bool) -> None:
     logger = logging.getLogger(__name__)
     if protect:
         msg = 'One or more intermediates of {} type need to be removed'
         logger.info(msg.format(','.join(intermediates)))
-        msg = 'Delete the "protect" flag in the config.json file or restart '\
+        msg = (
+            'Delete the "protect" flag in the config.json file or restart '
             'your analysis with another runid to continue.'
+        )
         logger.info(msg)
         end_program(1)
 
-    genome_tmpfiles = ['blast_results.tsv', 'keepsidx.json',
-                       'single_copy.json', 'blast_filtered.tsv',
-                       'expected_filt.json']
+    genome_tmpfiles = [
+        'blast_results.tsv',
+        'keepsidx.json',
+        'single_copy.json',
+        'blast_filtered.tsv',
+        'expected_filt.json',
+    ]
     if 'genome' in intermediates:
         for deldir in ['unaligned', 'aligned']:
             if os.path.exists(deldir):
                 shutil.rmtree(deldir)
-        for fullname in [os.path.join('.autoMLSA', fn) for
-                         fn in genome_tmpfiles]:
+        for fullname in [os.path.join('.autoMLSA', fn) for fn in genome_tmpfiles]:
             if os.path.exists(fullname):
                 os.remove(fullname)
     for delfile in glob.iglob(runid + '.nex*'):
-        os.remove(delfile)
+        logger.info('Backing up nexus and tree files to backup dir.')
+        shutil.move(delfile, 'backup')
+        if os.path.exists(delfile):
+            os.remove(delfile)
 
 
 def end_program(code: int) -> None:
